more reshuffling
