+++ examples/multi_no_thread.rs
@@ -1,9 +1,15 @@
+use inetnum::addr::Prefix;
 use log::trace;
 
+use rotonda_store::epoch;
 use rotonda_store::meta_examples::PrefixAs;
-use rotonda_store::prelude::multi::*;
-use rotonda_store::prelude::*;
-use rotonda_store::rib::MemoryOnlyConfig;
+use rotonda_store::IncludeHistory;
+use rotonda_store::IntoIpAddr;
+use rotonda_store::MatchOptions;
+use rotonda_store::MemoryOnlyConfig;
+use rotonda_store::Record;
+use rotonda_store::RouteStatus;
+use rotonda_store::StarCastRib;
 
 fn main() -> Result<(), Box<dyn std::error::Error>> {
     #[cfg(feature = "cli")]
@@ -11,7 +17,7 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
 
     trace!("Starting multi-threaded yolo testing....");
     let tree_bitmap =
-        MultiThreadedStore::<PrefixAs, MemoryOnlyConfig>::try_default()?;
+        StarCastRib::<PrefixAs, MemoryOnlyConfig>::try_default()?;
     // let f = Arc::new(std::sync::atomic::AtomicBool::new(false));
 
     let pfx = Prefix::new_relaxed(
