+++ examples/multi_thread_single_prefix.rs
@@ -1,7 +1,5 @@
 use log::trace;
 
-use rotonda_store::prelude::multi::*;
-use rotonda_store::rib::MemoryOnlyConfig;
 use std::sync::Arc;
 use std::thread;
 use std::time::Duration;
@@ -9,20 +7,20 @@ use std::time::Duration;
 use rand::Rng;
 
 use rotonda_store::meta_examples::PrefixAs;
-use rotonda_store::prelude::*;
-use rotonda_store::MultiThreadedStore;
+use rotonda_store::{
+    IncludeHistory, IntoIpAddr, MatchOptions, MemoryOnlyConfig, Record,
+    RouteStatus, StarCastRib,
+};
 
 fn main() -> Result<(), Box<dyn std::error::Error>> {
     #[cfg(feature = "cli")]
     env_logger::init();
 
     trace!("Starting multi-threaded yolo testing....");
-    let tree_bitmap = Arc::new(MultiThreadedStore::<
-        PrefixAs,
-        MemoryOnlyConfig,
-    >::try_default()?);
+    let tree_bitmap =
+        Arc::new(StarCastRib::<PrefixAs, MemoryOnlyConfig>::try_default()?);
 
-    let pfx = Prefix::new_relaxed(
+    let pfx = inetnum::addr::Prefix::new_relaxed(
         0b1111_1111_1111_1111_1111_1111_1111_1111_u32.into_ipaddr(),
         32,
     );
